more fixing vtk writing issue in surface meshing

+++ Source/Plugins/SurfaceMeshingPlugin/Code/SurfaceMeshing/SMVtkPolyDataWriter.h
@@ -135,11 +135,11 @@ class  SMVtkPolyDataWriter : public AbstractFilter
     SMVtkPolyDataWriter();
 
 
-    int writeBinaryCellData(const std::string &TrianglesFile, FILE* vtkFile, int nTriangles, bool conformalMesh);
-    int writeASCIICellData(const std::string &TrianglesFile, FILE* vtkFile, int nTriangles, bool conformalMesh);
+    int writeBinaryCellData(const std::string &TrianglesFile, FILE* vtkFile, size_t nTriangles, bool conformalMesh);
+    int writeASCIICellData(const std::string &TrianglesFile, FILE* vtkFile, size_t nTriangles, bool conformalMesh);
 
-    int writeBinaryPointData(const std::string &NodesFile, FILE* vtkFile, int nNodes, bool conformalMesh);
-    int writeASCIIPointData(const std::string &NodesFile, FILE* vtkFile, int nNodes, bool conformalMesh);
+    int writeBinaryPointData(const std::string &NodesFile, FILE* vtkFile, size_t nNodes, bool conformalMesh);
+    int writeASCIIPointData(const std::string &NodesFile, FILE* vtkFile, size_t nNodes, bool conformalMesh);
 
   private:
     SMVtkPolyDataWriter(const SMVtkPolyDataWriter&); // Copy Constructor Not Implemented
